# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey API",
        "description": "REST API сервиса опросов и голосований",
        "contact": {
            "name": "API Support",
            "email": "support@survey.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/surveys": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Список опросов с пагинацией и фильтрами",
                "operationId": "7e96246f34de1fcd663f2d95a5079073",
                "parameters": [
                    {
                        "name": "mine",
                        "in": "query",
                        "description": "Только мои опросы",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "published",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Поле сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "responses_count"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Направление",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Страница",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список опросов"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{survey}": {
            "put": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Обновить опрос (только черновик и автор)",
                "operationId": "004b5441c458eb1e748628dff71f780f",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Обновленное название"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Новое описание"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Опрос обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Survey"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Это не ваш опрос"
                    },
                    "422": {
                        "description": "Редактировать можно только черновик"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Удалить опрос (только черновик и автор)",
                "operationId": "56fe63d4026da0b796e6ff82b49cf12f",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Опрос удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Опрос удалён"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Это не ваш опрос"
                    },
                    "422": {
                        "description": "Редактировать можно только черновик"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Survey": {
                "title": "Опрос",
                "description": "Опрос с вопросами",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID автора",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Опрос о языках программирования"
                    },
                    "description": {
                        "type": "string",
                        "example": "Выберите ваш любимый язык"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "closed"
                        ],
                        "example": "draft"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "questions": {
                        "description": "Вопросы опроса",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "type": "object"
            },
            "Question": {
                "title": "Вопрос",
                "description": "Вопрос в опросе",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "survey_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Какой ваш любимый язык?"
                    },
                    "type": {
                        "description": "Тип вопроса: single (одиночный выбор), multiple (множественный выбор), text (текстовый)",
                        "type": "string",
                        "enum": [
                            "single",
                            "multiple",
                            "text"
                        ],
                        "example": "single"
                    },
                    "order": {
                        "description": "Порядок вопроса",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "options": {
                        "description": "Варианты ответов (только для single/multiple)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        }
                    }
                },
                "type": "object"
            },
            "Option": {
                "title": "Вариант ответа",
                "description": "Вариант ответа на вопрос",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "PHP"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT токен в формате: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}